requirements.txt
# GCP SDKs
google-cloud-aiplatform==1.60.0
google-cloud-bigquery==3.25.0
google-cloud-storage==2.18.0
google-cloud-monitoring==2.22.0
google-cloud-run==0.10.0
google-cloud-secret-manager==2.20.0
google-cloud-logging==3.11.0

# Vertex AI Pipelines (Kubeflow)
kfp==2.7.0
google-cloud-pipeline-components==2.15.0

# ML Frameworks
scikit-learn==1.5.0
xgboost==2.1.0
lightgbm==4.4.0
pandas==2.2.0
numpy==1.26.4

# API Serving
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0

# Data Validation
great-expectations==0.18.0
pandera==0.20.0

# Monitoring & Drift
evidently==0.4.33
scipy==1.14.0

# Experiment Tracking
mlflow==2.16.0

# Utils
python-dotenv==1.0.1
loguru==0.7.2
pyyaml==6.0.1
click==8.1.7
httpx==0.27.0
joblib==1.4.0

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==5.0.0

# Code Quality
black==24.8.0
ruff==0.6.0
mypy==1.11.0
